test: move redundant config to shared test utils
