+++ test/utils/index.js
@@ -7,3 +7,4 @@ exports.connectPeers = require('./connect-peers')
 exports.MemStore = require('./mem-store')
 exports.CustomTestKeystore = require('./custom-test-keystore')
 exports.CustomTestCache = require('./custom-test-cache')
+exports.databases = require('./databases')
